JPA repository, dependency injection, database seeding

+++ src/main/java/com/educandoweb/course/entities/User.java
@@ -12,13 +12,13 @@ public class User {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
+  private Long id;
   private String name;
   private String email;
   private String phone;
   private String password;
 
-  public User(Integer id, String name, String email, String phone, String password) {
+  public User(Long id, String name, String email, String phone, String password) {
     this.id = id;
     this.name = name;
     this.email = email;
@@ -29,11 +29,11 @@ public class User {
   public User() {
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
